+++ ThirdPartyAdapters/unity/unity/src/main/java/com/google/ads/mediation/unity/UnityMediationBannerAd.java
@@ -39,7 +39,6 @@ import com.google.android.gms.ads.mediation.MediationBannerAdConfiguration;
 import com.unity3d.ads.IUnityAdsInitializationListener;
 import com.unity3d.ads.UnityAds;
 import com.unity3d.ads.UnityAdsLoadOptions;
-import com.unity3d.ads.metadata.MetaData;
 import com.unity3d.services.banners.BannerErrorInfo;
 import com.unity3d.services.banners.BannerView;
 import com.unity3d.services.banners.UnityBannerSize;
@@ -204,8 +203,7 @@ public class UnityMediationBannerAd implements MediationBannerAd, BannerView.ILi
                     gameId, bannerPlacementId);
             Log.d(UnityMediationAdapter.TAG, logMessage);
 
-            UnityAdsAdapterUtils.setUnityAdsPrivacy(
-                MobileAds.getRequestConfiguration(), new MetaData(context));
+            UnityAdsAdapterUtils.setUnityAdsPrivacy(MobileAds.getRequestConfiguration());
 
             if (unityBannerViewWrapper == null) {
               unityBannerViewWrapper =